Support for clearing the cone if on max reading
This change is to add a parameter that allows for the clearing of all elements in the sensor cone in the event of a range reading above the max value in the range message.  In doing so, the map can keep clear of ghost obstacles in the case where a sensor reading is given above max range for no detected obstacles within range.  This behavior matches better with the semantics of a reading above max range.  It is an additional parameter for the layer though, in case this behavior isn't present in the sensor providing the range messages.